[#128] Refactor test suites
