+++ tests/campusToggle.test.js
@@ -1,8 +1,9 @@
-/* eslint-disable no-undef */
 /* eslint-disable max-len */
 import React from 'react';
 import renderer from 'react-test-renderer';
 import CampusToggle from '../components/campusToggle';
+
+let updateR;
 
 beforeEach(() => {
   updateR = jest.fn();
