+++ tests/destinationSearchBar.test.js
@@ -1,8 +1,11 @@
-/* eslint-disable no-undef */
 /* eslint-disable max-len */
 import React from 'react';
 import renderer from 'react-test-renderer';
-import DestinationSearchBar from '../components/destinationSearchBar';
+import DestinationSearchBar from '../components/directions/destinationSearchBar';
+
+let mockPrediction;
+let mockResult;
+let updateR;
 
 beforeEach(() => {
   mockPrediction = {
